+++ tests/test_httpx_async.py
@@ -609,12 +609,15 @@ async def test_requests_json_body(httpx_mock: HTTPXMock):
     async with httpx.AsyncClient() as client:
         response = await client.post("http://test_url")
         assert response.json() == {"key 1": "value 1", "key 2": "value 2"}
+        assert response.headers["content-type"] == "application/json"
 
         response = await client.get("http://test_url")
         assert response.json() == ["list content 1", "list content 2"]
+        assert response.headers["content-type"] == "application/json"
 
         response = await client.put("http://test_url")
         assert response.json() == "string value"
+        assert response.headers["content-type"] == "application/json"
 
 
 @pytest.mark.asyncio
@@ -642,6 +645,7 @@ async def test_callback_returning_response(httpx_mock: HTTPXMock):
     async with httpx.AsyncClient() as client:
         response = await client.get("http://test_url")
         assert response.json() == {"url": "http://test_url"}
+        assert response.headers["content-type"] == "application/json"
 
 
 @pytest.mark.asyncio
@@ -654,9 +658,11 @@ async def test_callback_executed_twice(httpx_mock: HTTPXMock):
     async with httpx.AsyncClient() as client:
         response = await client.get("http://test_url")
         assert response.json() == ["content"]
+        assert response.headers["content-type"] == "application/json"
 
         response = await client.post("http://test_url")
         assert response.json() == ["content"]
+        assert response.headers["content-type"] == "application/json"
 
 
 @pytest.mark.asyncio
@@ -669,9 +675,11 @@ async def test_callback_matching_method(httpx_mock: HTTPXMock):
     async with httpx.AsyncClient() as client:
         response = await client.get("http://test_url")
         assert response.json() == ["content"]
+        assert response.headers["content-type"] == "application/json"
 
         response = await client.get("http://test_url2")
         assert response.json() == ["content"]
+        assert response.headers["content-type"] == "application/json"
 
 
 def test_request_retrieval_with_more_than_one(testdir):